add focus: true to Container.WithExec
Focused execs will be featured more prominently in the UI. You would
typically use this flag to mark the command that accomplishes the
primary goal of the pipeline, such as the command that runs your tests.
