+++ engine/engine.go
@@ -36,7 +36,6 @@ import (
 	fstypes "github.com/tonistiigi/fsutil/types"
 	"github.com/vito/progrock"
 	"golang.org/x/sync/errgroup"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 const (
@@ -207,12 +206,7 @@ func Start(ctx context.Context, startOpts Config, fn StartCallback) error {
 	eg, groupCtx := errgroup.WithContext(ctx)
 	solveCh := make(chan *bkclient.SolveStatus)
 	eg.Go(func() error {
-		for ev := range solveCh {
-			if err := recorder.Record(bk2progrock(ev)); err != nil {
-				return fmt.Errorf("record: %w", err)
-			}
-		}
-		return nil
+		return core.RecordBuildkitStatus(recorder, solveCh)
 	})
 
 	eg.Go(func() error {
@@ -355,66 +349,6 @@ func cacheConfigFromEnv() (string, map[string]string, error) {
 	}
 	delete(attrs, "type")
 	return typeVal, attrs, nil
-}
-
-func bk2progrock(event *bkclient.SolveStatus) *progrock.StatusUpdate {
-	var status progrock.StatusUpdate
-	for _, v := range event.Vertexes {
-		vtx := &progrock.Vertex{
-			Id:     v.Digest.String(),
-			Name:   v.Name,
-			Cached: v.Cached,
-		}
-		if strings.HasPrefix(v.Name, "[internal] ") {
-			vtx.Internal = true
-			vtx.Name = strings.TrimPrefix(v.Name, "[internal] ")
-		}
-		for _, input := range v.Inputs {
-			vtx.Inputs = append(vtx.Inputs, input.String())
-		}
-		if v.Started != nil {
-			vtx.Started = timestamppb.New(*v.Started)
-		}
-		if v.Completed != nil {
-			vtx.Completed = timestamppb.New(*v.Completed)
-		}
-		if v.Error != "" {
-			if strings.HasSuffix(v.Error, context.Canceled.Error()) {
-				vtx.Canceled = true
-			} else {
-				msg := v.Error
-				vtx.Error = &msg
-			}
-		}
-		status.Vertexes = append(status.Vertexes, vtx)
-	}
-
-	for _, s := range event.Statuses {
-		task := &progrock.VertexTask{
-			Vertex:  s.Vertex.String(),
-			Name:    s.ID, // remap
-			Total:   s.Total,
-			Current: s.Current,
-		}
-		if s.Started != nil {
-			task.Started = timestamppb.New(*s.Started)
-		}
-		if s.Completed != nil {
-			task.Completed = timestamppb.New(*s.Completed)
-		}
-		status.Tasks = append(status.Tasks, task)
-	}
-
-	for _, s := range event.Logs {
-		status.Logs = append(status.Logs, &progrock.VertexLog{
-			Vertex:    s.Vertex.String(),
-			Stream:    progrock.LogStream(s.Stream),
-			Data:      s.Data,
-			Timestamp: timestamppb.New(s.Timestamp),
-		})
-	}
-
-	return &status
 }
 
 func progrockForwarder(w progrock.Writer) (string, progrock.Writer, func() error, error) {
